% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{lcvp_authors}
\alias{lcvp_authors}
\title{List of authors and their position in tab_lcvp}
\format{
A list.
}
\usage{
lcvp_authors
}
\description{
The 'lcvp_authors' list includes all authors and their respective
described species' position in the tab_lcvp.
}
\examples{

data(lcvp_authors)

}
\keyword{datasets}
