% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{lcvp_family}
\alias{lcvp_family}
\title{List of families and their position in tab_lcvp}
\format{
A list.
}
\usage{
lcvp_family
}
\description{
The 'lcvp_family' list includes all families and their respective
position in the tab_lcvp.
}
\examples{

data(lcvp_family)

}
\keyword{datasets}
