% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{lcvp_order}
\alias{lcvp_order}
\title{List of orders and their position in tab_lcvp}
\format{
A list.
}
\usage{
lcvp_order
}
\description{
The 'lcvp_order' list includes all families and their respective
position in the tab_lcvp.
}
\examples{

data(lcvp_order)

}
\keyword{datasets}
