% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{lcvp_sps_class}
\alias{lcvp_sps_class}
\title{List of species name in tab_lcvp separeted by category}
\format{
A data.frame.
}
\usage{
lcvp_sps_class
}
\description{
The 'lcvp_sps_class' includes all species separeted by genus, epithet, author, 
subspecies, variety, subvariety, forma, subforma and ID (position in the 
\code{tab_lcvp}).
position in the \code{tab_lcvp}.
}
\examples{

data(lcvp_sps_class)

}
\keyword{datasets}
