% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{match_lcvp}
\alias{match_lcvp}
\title{Link between Taxa and Literature Reference}
\format{
A data frame with 104,860 observations on the following 2 variables.
}
\description{
The 'ref_lcvp' links all taxa in LCVP with the relevant literature from ref_lcvp
}
\examples{

data(match_lcvp)
head(match_lcvp)

}
\keyword{datasets}
