% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{ref_lcvp}
\alias{ref_lcvp}
\title{Reference List of LCVP}
\format{
A data frame with 4484 observations on the following 16 variables.
}
\description{
The 'ref_lcvp' contains the references to all literature used for compiling LCVP
}
\examples{

data(ref_lcvp)
str(ref_lcvp)

}
\references{
The Leipzig Catalogue of Vascular Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants
}
\keyword{datasets}
