% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{tab_lcvp}
\alias{tab_lcvp}
\title{List of Plant Species Name accordingly with the Leipzig Catalogue of Vascular Plants
(LCVP)}
\format{
A data frame with the following columns:
  \describe{\item{global.Id}{The fixed species id of the input taxon in the
  Leipzig Catalogue of Vascular Plants (LCVP).} 
  \item{Input.Genus}{A
  character vector. The input genus of the corresponding vascular plant
  species name listed in LCVP.} 
  \item{Input.Epitheton}{A character vector.
  The input epitheton of the corresponding vascular plant species name listed
  in LCVP.} 
  \item{Rank}{A character vector. The taxonomic rank ("species",
  subspecies: "subsp.", variety: "var.", subvariety: "subvar.", "forma", or
  subforma: "subf.") of the corresponding vascular plant species name listed
  in LCVP.} 
  \item{Input.Subspecies.Epitheton}{A character vector. If the
  indicated rank is below species, the subspecies epitheton input of the
  corresponding vascular plant species name listed in LCVP. If the rank is
  "species", the input is "nil".} 
  \item{Input.Authors}{A character vector.
  The taxonomic authority input of the corresponding vascular plant species
  name listed in LCVP.} 
  \item{Status}{A character vector. description if a
  taxon is classified as ‘valid’, ‘synonym’, ‘unresolved’, ‘external’ or
  ‘blanks’. The ‘unresolved’ rank means that the status of the plant name
  could be either valid or synonym, but the information available does not
  allow a definitive decision. ‘External’ in an extra rank which lists names
  outside the scope of this publication but useful to keep on this updated
  list. ‘Blanks’ means that the respective name exists in bibliography but it
  is neither clear where it came from valid, synonym or unresolved. (see the
  main text Freiberg et al. for more details)}
  \item{globalId.of.Output.Taxon}{The fixed species id of the output taxon
  in LCVP.} 
  \item{Output.Taxon}{A character vector. The list of the accepted
  plant taxa names according to the LCVP.} 
  \item{Family}{A character vector.
  The corresponding family name of the Input.Taxon, staying empty if the
  Status is unresolved.} 
  \item{Order}{A character vector. The corresponding
  order name of the Input.Taxon, staying empty if the Status is unresolved.}
  \item{Literature}{A character vector. The bibliography used.}
  \item{Comments}{A character vector. Further taxonomic comments.}}
}
\usage{
tab_lcvp
}
\description{
The 'tab_lcvp' contains records belonging to all the
vascular plant taxa names checked and listed from the "Leipzig Catalogue of
Plants" (LCVP).
}
\examples{

data(tab_lcvp)
str(tab_lcvp)

}
\references{
The Leipzig Catalogue of Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants.
}
\keyword{datasets}
