% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_LCVP_data.R
\name{update_LCVP_data}
\alias{update_LCVP_data}
\title{Update the data files of LCVP
Updates the tab-lcsv and tab_position data files from a new text file version of the Leipzig CAtalogue of Vascular Plants.}
\usage{
update_LCVP_data(
  input_data = "raw_data_LCVP/LCVP_103.txt",
  output_folder = "data",
  output_names = c("tab_lcvp.rda", "tab_position.rda"),
  encoding = "ANSI"
)
}
\arguments{
\item{input_data}{the path to the new version of the Leipzig Catalogue of Vasculart plants.}

\item{output_folder}{the floder where the output files should be saved. By default the data 
folder of the LCVP package.}

\item{output_names}{The names of the output files. Position one is the tab_lcsvp files
position two the tab_position file.}

\item{encoding}{the encoding to read the input data.}
}
\description{
Update the data files of LCVP
Updates the tab-lcsv and tab_position data files from a new text file version of the Leipzig CAtalogue of Vascular Plants.
}
\examples{

\dontrun{
update_LCVP_data()
}

}
\author{
Alessandro Gentile, Alexander Zizka
}
