% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{tab_lcvp}
\alias{tab_lcvp}
\title{List of Plant Species Name accordingly with the Leipzig Catalogue of Vascular Plants
(LCVP)}
\format{
A data frame with observations on the following 7 variables.
\describe{ \item{Input.Taxon}{A character vector. This variable
contains the list of all 1.314.778 vascular plant taxa names listed by the
authors of the Leipzig Catalogue of Plants.} \item{Status}{A
character vector. This variable provides a description if a taxon is
classified as 'valid', 'synonym', 'unresolved' or 'external'. The
'unresolved' rank means that the status of the plant name could be either
valid or synonym, but the information available does not allow a definitive
decision (see the main text Freiberg et al. for more details).}
\item{PL.comparison}{A character vector. This field provides a
direct comparison with 'The Plant List' (TPL; The Plant List
http://www.theplantlist.org/ accessed: 1.1. 2013) reporting further
information such as 'identical', 'synonym', 'other synonym', 'different
authors', 'missing', 'misspelling', 'unresolved'.}
\item{PL.alternative}{A character vector. This field provides a
possible alternative name suggested by TPL.} \item{Output.Taxon}{A
character vector. This variable contains the list of the accepted plant taxa
names (350.766) according to the LCVP.} \item{Family}{A character
vector. It provides the corresponding family name of the 'Input.Taxon',
staying empty if the Status is unresolved.} \item{Order}{A character
vector. The Order column provides the corresponding order name of the
'Input.Taxon', staying empty if the Status is unresolved.} }
}
\source{
https://idata.idiv.de/ddm/Data/ShowData/1806
}
\description{
The 'tab_lcvp' contains records belonging to all the
vascular plant taxa names checked and listed from the "Leipzig Catalogue of
Plants" (LCVP).
}
\examples{

data(tab_lcvp)
str(tab_lcvp)

}
\references{
The Leipzig Catalogue of Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants.
}
\keyword{datasets}
