% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{tab_lcvp}
\alias{tab_lcvp}
\title{List of Plant Species Name accordingly with the Leipzig Catalogue of Vascular Plants
(LCVP)}
\format{
A data frame with observations on the following 7 variables.
\describe{ \item{Input.Taxon}{A character vector. 
The list of all vascular plant species names listed in the Leipzig Catalogue of Vascular Plants (LCVP).} 
\item{Status}{A
character vector. description if a taxon is classified as 
‘valid’, ‘synonym’, ‘unresolved’, ‘external’ or ‘blanks’.
The ‘unresolved’ rank means that the status of the plant 
name could be either valid or synonym, but the information 
available does not allow a definitive decision.
‘External’ in an extra rank which lists names outside the 
scope of this publication but useful to keep on this updated list.
‘Blanks’ means that the respective name exists in bibliography 
but it is neither clear where it came from valid, synonym or unresolved.
(see the main text Freiberg et al. for more details).}
\item{PL.comparison}{A character vector. This field provides a direct 
comparison with ‘The Plant List’ 
(TPL; The Plant List \url{http://www.theplantlist.org/} accessed: 1.1. 2013) 
reporting further information such as ‘identical', 'synonym', 
'other synonym', 'different authors', 'missing', 'misspelling', 'unresolved'.}
\item{PL.alternative}{A character vector. This field provides a
possible alternative name from the TPL.} 
\item{Output.Taxon}{A
character vector. The list of the accepted plant 
taxa names according to the LCVP.}
\item{Family}{A character
vector. The corresponding family name of the Input.Taxon, 
staying empty if the Status is unresolved..} 
\item{Order}{A character
vector. The corresponding order name of the Input.Taxon, 
staying empty if the Status is unresolved.} }
}
\usage{
tab_lcvp
}
\description{
The 'tab_lcvp' contains records belonging to all the
vascular plant taxa names checked and listed from the "Leipzig Catalogue of
Plants" (LCVP).
}
\examples{

data(tab_lcvp)
str(tab_lcvp)

}
\references{
The Leipzig Catalogue of Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants.
}
\keyword{datasets}
