% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LCVP-package.R
\docType{data}
\name{tab_position}
\alias{tab_position}
\title{List of the number positions of the first 3 letters of the species name in
the tab_lcvp}
\format{
A data frame with 2676 observations on the following 3 variables.
\describe{ 
 \item{Position}{A character vector. It is the position of
the first 3 letters of the species name in the tab_lcvp.}
\item{Triphthong}{A character vector. First 3 letters of the species
name in the tab_lcvp.} \item{Genus}{A character vector.
Corresponding Genus name.} }
}
\usage{
tab_position
}
\description{
The 'tab_position' reports the
position (in term of number of rows) of the first three letters (triphthong)
for the plant names stored in the variable 'Input.Taxon' (which
are listed in alphabetical order) of the table 'tab_lcvp'. This
indexing system speeds up of the search on the largest list
'tab_lcvp.rda' using the lcvplants package.
}
\examples{

data(tab_position)
str(tab_position)
}
\references{
The Leipzig Catalogue of Vascular Plants (LCVP) - An improved taxonomic
reference list for all known vascular plants
}
\keyword{datasets}
